[feature/admin]internPJのadminを実装 (#1)

+++ app/migrations/20220207082946_create_admins.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE admins (
-  id         bigint(20) NOT NULL AUTO_INCREMENT,
+  id         INT NOT NULL AUTO_INCREMENT,
   name       varchar(255) NOT NULL UNIQUE,
   email      varchar(255) NOT NULL UNIQUE,
   password   varchar(255) NOT NULL,
